scripts/requirements.txt
progressbar2
spacy
gensim
sklearn